modularized `pbl` and some basic badge design work
